# openapi.yaml
#
# Copyright 2021 The Sigstore Authors.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

swagger: '2.0'
info:
  version: '1.0.0'
  title: Fulcio
schemes: [http, https]
host: fulcio.sigstore.dev
basePath: /
securityDefinitions:
  Bearer:
    type: apiKey
    in: header
    name: Authorization
paths:
  /:
    get:
      description: 'open HomePage'
      operationId: homepage
      produces:
        - text/plain
      responses:
        200:
          $ref: '#/responses/OK'
        400:
          $ref: '#/responses/BadRequest'
        default:
          $ref: '#/responses/InternalServerError'
  /api/v1/signingCert:
    post:
      security:
        - Bearer: []
      description: 'create a cert, return content with a location header (with URL to CTL entry)'
      operationId: signingCert
      consumes:
        - application/json
      produces:
        - application/pem-certificate-chain
      parameters:
        - name: 'CertificateRequest'
          in: 'body'
          required: true
          description: 'Request for signing certificate'
          schema:
            $ref: '#/definitions/CertificateRequest'
      responses:
        201:
          description: Generated Certificate Chain
          schema:
            type: string
        400:
          $ref: '#/responses/BadRequest'
        401:
          $ref: '#/responses/Unauthorized'
        default:
          $ref: '#/responses/InternalServerError'

definitions:
  CertificateRequest:
    type: object
    properties:
      publicKey:
        type: object
        properties:
          content:
            type: string
            format: byte
          algorithm:
            type: string
            enum: ['ecdsa']
        required:
          - content
      signedEmailAddress:
        type: string
        format: byte
    required:
      - publicKey
      - signedEmailAddress
  Error:
    type: object
    properties:
      code:
        type: integer
      message:
        type: string

responses:
  OK:
    description: Ok message
    schema:
      type: string
    headers:
      Content-Type:
        type: string
  BadRequest:
    description: The content supplied to the server was invalid
    schema:
      $ref: "#/definitions/Error"
    headers:
      Content-Type:
        type: string
  Unauthorized:
    description: The request could not be authorized
    schema:
      $ref: "#/definitions/Error"
    headers:
      Content-Type:
        type: string
      WWW-Authenticate:
        type: string
        description: Information about required authentication to access server
  InternalServerError:
    description: There was an internal error in the server while processing the request
    schema:
      $ref: "#/definitions/Error"
    headers:
      Content-Type:
        type: string